+++ include/eve/module/math/detail/constant/rempio2_limits.hpp
@@ -7,16 +7,12 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/detail/overload.hpp>
-#include <eve/detail/abi.hpp>
-#include <eve/function/trigo_tags.hpp>
-#include <eve/constant/ieee_constant.hpp>
-#include <eve/constant/valmax.hpp>
-#include <eve/constant/pio_4.hpp>
-#include <eve/constant/pio_2.hpp>
-#include <eve/constant/pi.hpp>
-#include <eve/concept/value.hpp>
-#include <eve/as.hpp>
+#include <eve/module/ieee.hpp>
+#include <eve/module/math/regular/trigo_tags.hpp>
+#include <eve/module/math/constant/pio_4.hpp>
+#include <eve/module/math/constant/pio_2.hpp>
+#include <eve/module/math/constant/pi.hpp>
+
 
 namespace eve::detail
 {
