+++ include/eve/module/math/detail/generic/rem2.hpp
@@ -7,17 +7,10 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/concept/value.hpp>
-#include <eve/constant/mhalf.hpp>
-#include <eve/constant/pi.hpp>
-#include <eve/detail/abi.hpp>
+#include <eve/module/arithmetic.hpp>
 #include <eve/detail/kumi.hpp>
-#include <eve/detail/meta.hpp>
-#include <eve/detail/overload.hpp>
-#include <eve/function/fma.hpp>
-#include <eve/function/nearest.hpp>
-#include <eve/function/quadrant.hpp>
-#include <eve/function/two_prod.hpp>
+#include <eve/module/math/constant/pi.hpp>
+#include <eve/module/math/regular/quadrant.hpp>
 
 namespace eve::detail
 {
@@ -31,4 +24,3 @@ namespace eve::detail
     return kumi::make_tuple(quadrant(xi), xr, dxr);
   }
 }
-
